Fix documentation formatting, minor fixes
