+++ engine/launcher/src/test/scala/org/enso/launcher/NativeTest.scala
@@ -1,11 +1,24 @@
 package org.enso.launcher
 
-import org.scalatest.Suite
+import org.scalatest.matchers.should.Matchers
 import org.scalatest.matchers.{MatchResult, Matcher}
+import org.scalatest.wordspec.AnyWordSpec
 
 import scala.sys.process._
 
-trait NativeTest extends Suite {
+/**
+  * Contains helper methods for creating tests that need to run the native
+  * launcher binary.
+  */
+trait NativeTest extends AnyWordSpec with Matchers {
+
+  /**
+    * A result of running the native launcher binary.
+    *
+    * @param exitCode the returned exit code
+    * @param stdout contents of the standard output stream
+    * @param stderr contents of the standard error stream
+    */
   case class RunResult(exitCode: Int, stdout: String, stderr: String)
 
   object RunSuccessMatcher extends Matcher[RunResult] {
@@ -17,8 +30,21 @@ trait NativeTest extends Suite {
       )
   }
 
-  def returnSuccess = RunSuccessMatcher
+  /**
+    * A scalatest Matcher that allows to write expressions like
+    * `run should returnSuccess`.
+    */
+  def returnSuccess: RunSuccessMatcher.type = RunSuccessMatcher
 
+  /**
+    * Runs the native launcher binary with the specified arguments and captures
+    * its output.
+    *
+    * @param args arguments to forward to the launcher
+    * @param extraEnv additional environment variables that should be overriden
+    *                 for the launcher binary
+    * @return
+    */
   def runLauncher(
     args: Seq[String],
     extraEnv: (String, String)*
